Formerly ../posix/getconf.c.~4~

+++ posix/getconf.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1991 Free Software Foundation, Inc.
+/* Copyright (C) 1991, 1992 Free Software Foundation, Inc.
 This file is part of the GNU C Library.
 
 The GNU C Library is free software; you can redistribute it and/or
@@ -124,7 +124,7 @@ DEFUN(main, (argc, argv), int argc AND char **argv)
 			program, strerror(errno));
 		exit(3);
 	      }
-	    printf("%.*s\n", clen, cvalue);
+	    printf("%.*s\n", (int) clen, cvalue);
 	    exit(0);
 	  }
       }
